Using primitive data type 'byte' for bytes (iso short)

+++ src/org/maia/amstrad/pc/basic/BasicDecompiler.java
@@ -2,6 +2,6 @@ package org.maia.amstrad.pc.basic;
 
 public interface BasicDecompiler {
 
-	CharSequence decompile(short[] byteCode);
+	CharSequence decompile(byte[] byteCode);
 
 }
